modularized collector.cpp, added header files for info structure and ifconfig parser
